Rename volume3d to volume
