+++ skimage/measure/__init__.py
@@ -15,7 +15,7 @@ from .fit import LineModelND, CircleModel, EllipseModel, ransac
 from .block import block_reduce
 from ._label import label
 from .entropy import shannon_entropy
-from .volume3d import expanded_convex_hull
+from ._volume import expanded_convex_hull
 
 
 __all__ = ['find_contours',
