+++ skimage/measure/_volume.py
